// src/logic/rest/api.ts
export interface Body7 {
    /**
     * Id of the User
     * @type {number}
     * @memberof Body7
     */
    userId?: number;
}
// ...
export interface Room {
    /**
     * 
     * @type {number}
     * @memberof Room
     */
    id: number;
    /**
     * 
     * @type {string}
     * @memberof Room
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Room
     */
    description: string;
    /**
     * 
     * @type {Date}
     * @memberof Room
     */
    createdon: Date;
    /**
     * 
     * @type {boolean}
     * @memberof Room
     */
    private: boolean;
    /**
     * 
     * @type {number}
     * @memberof Room
     */
    ownerId: number;
    /**
     * 
     * @type {string}
     * @memberof Room
     */
    ownerName: string;
}